+++ src/pages/calendar/Calendar.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react'
 import Modal from 'react-modal'
-import CustomButton from '../../components/customButton/customButton'
-import InputField from '../../components/inputField/InputField'
+import CustomButton from '../../components/CustomButton/CustomButton'
+import InputField from '../../components/InputField/InputField'
 import { addEvent, listenToEvents } from '../../services/firestoreService'
 import { useUser } from '../../utils/hooks'
 import './Calendar.css'
